TopicモデルのPHPDocと型定義の見直し

+++ app/Http/Controllers/CommentController.php
@@ -71,7 +71,7 @@ class CommentController extends Controller
         $input = $request->all();
 
         /* トピックの存在確認 */
-        $topic = $this->m_topic->getTopicById(Arr::get($input, 'topic_id'));
+        $topic = $this->m_topic->getTopicById((int) Arr::get($input, 'topic_id'));
         if (!isset($topic)) {
             /* 不正なIDもしくはトピックが存在しない場合は一覧に戻す */
             session()->flash('flash_failed', __('comments.fail.not_exist'));
